Rename feature.rs file to feature_snapshot.rs, to match the struct name
