+++ src/client/feature_snapshot.rs
